% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmd_lqcoeffs}
\alias{hmd_lqcoeffs}
\title{Log-Quadratic Model coefficients for Human Mortality Database}
\format{
A list with 3 elements representing coefficients for both sexes (total), males (male) and females (female)
}
\usage{
hmd_lqcoeffs
}
\description{
Log-Quadratic Model coefficients for Human Mortality Database
}
\keyword{datasets}
