% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WPP2019_asfr}
\alias{WPP2019_asfr}
\title{ASFR output of WPP2019, including all populations. This dataset is temporary, for convenience in \code{DemoTools} and \code{fertestr} and it may be deprecated when better estimates become available.}
\format{
A \code{data.frame} with 5 columns and 24990 rows
\itemize{
\item \code{LocID} character. UN location codes numeric, but stored as character.
\item \code{Year} integer
\item \code{AgeStart} integer lower bound of the age interval.
\item \code{AgeSpan} integer age interval width.
\item \code{ASFR} double. Age specific fertility rate on a per-woman scale.
}
}
\usage{
WPP2019_asfr
}
\description{
ASFR output of WPP2019, including all populations. This dataset is temporary, for convenience in \code{DemoTools} and \code{fertestr} and it may be deprecated when better estimates become available.
}
\keyword{datasets}
