% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WPP2019_lt}
\alias{WPP2019_lt}
\title{Selected lifetable output of WPP2019, including all populations. This dataset is temporary, for convenience in \code{DemoTools} and \code{fertestr} and it may be deprecated when better estimates become available. This data may need to be interpolated to. Total population by age can be calculated as the sum of male and female populations.}
\format{
A \code{data.frame} with 7 columns and 235620 rows
\itemize{
\item \code{LocID} character. UN location codes numeric, but stored as character.
\item \code{Year} integer.
\item \code{Sex} character. \code{"b"}, \code{"f"}, or \code{"m"}
\item \code{AgeStart} integer lower bound of the abridged age interval.
\item \code{AgeSpan} integer age interval width.
\item \code{mx} double. death rate.
\item \code{lx} double. radix 100000 survivorship.
}
}
\usage{
WPP2019_lt
}
\description{
Selected lifetable output of WPP2019, including all populations. This dataset is temporary, for convenience in \code{DemoTools} and \code{fertestr} and it may be deprecated when better estimates become available. This data may need to be interpolated to. Total population by age can be calculated as the sum of male and female populations.
}
\keyword{datasets}
