% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WPP_codes}
\alias{WPP_codes}
\title{A convenience lookup table for location codes. Some WPP data objects (\code{WPP2019_lt}, \code{WPP2019_pop}, \code{WPP2019_asfr}) only have the UN \code{LocID} code. The object may help with more flexible queries.}
\format{
A \code{data.frame} with 4 columns and 235 rows
\itemize{
\item \code{LocID} character. UN location codes numeric, but stored as character.
\item \code{ISO3} character. 3-letter location codes.
\item \code{ISO2} character. 2-letter location codes.
\item \code{LocName} character. Spelled-out location name.
}
}
\usage{
WPP_codes
}
\description{
A convenience lookup table for location codes. Some WPP data objects (\code{WPP2019_lt}, \code{WPP2019_pop}, \code{WPP2019_asfr}) only have the UN \code{LocID} code. The object may help with more flexible queries.
}
\keyword{datasets}
